+++ packages/luis/src/commands/luis/test.ts
@@ -8,12 +8,12 @@ import {CognitiveServicesCredentials} from '@azure/ms-rest-azure-js'
 import {CLIError, Command, flags, utils} from '@microsoft/bf-cli-command'
 const fs = require('fs-extra')
 const file = require('@microsoft/bf-lu/lib/utils/filehelper')
-const testHelper = require('@microsoft/bf-lu/lib/utils/testhelper')
+const testHelper = require('@microsoft/bf-lu/lib/parser/test/testhelper')
+const testLuConverter = require('@microsoft/bf-lu/lib/parser/test/testLuConverter')
 const exception = require('@microsoft/bf-lu').V2.Exception
 const fileExtEnum = require('@microsoft/bf-lu/lib/parser/utils/helpers').FileExtTypeEnum
 const LuisBuilder = require('@microsoft/bf-lu/lib/parser/luis/luisCollate')
 const Luis = require('@microsoft/bf-lu').V2.Luis
-const luConverterForTest = require('./../../../../lu/src/parser/test/luConverterForTest')
 
 import {hasContent} from './../../utils/luisinstanceutils'
 export default class LuisTest extends Command {
@@ -79,7 +79,7 @@ export default class LuisTest extends Command {
         flags.intentOnly,
         predictedResults)
 
-      let result = luConverterForTest(luisObject, flags)
+      let result = testLuConverter(luisObject, flags)
       let detailLog = `${JSON.stringify(predictedResults, null, 2)}`
       if (!result) {
         throw new CLIError('No LU content parsed!')
